refactored SignUpForm to use the fetch api instead of axios due to testing issues
