+++ frontend/src/components/SignUpForm/SignUpForm.css
@@ -1,0 +1,6 @@
+.success-message{
+    color: green;
+    text-align: center;
+    font-size: large;
+    font-weight: bold;
+}
