+++ frontend/src/components/SignUpForm/SignUpForm.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import SignUp from './SignUpForm'
+
+test('Sign up page is rendered', () => {
+  render(<SignUp />);
+  const linkElement = screen.getByText(/sign up/i);
+  expect(linkElement).toBeInTheDocument();
+});
