+++ frontend/src/components/globalStyle.css
@@ -99,9 +99,3 @@ button:hover {
     color: black;
 }
 
-.success-message{
-    color: green;
-    text-align: center;
-    font-size: large;
-    font-weight: bold;
-}
